refactor(tests): move data_consistency in batch tests.
